+++ un2projeto/lib/componentes/tarefa_item.dart
@@ -23,9 +23,10 @@ class TarefaItem extends StatelessWidget {
           subtitle: Text("Usuário: " + tarefa.usuario.nome),
           shape: RoundedRectangleBorder(
               side: const BorderSide(
-                  color: Color.fromARGB(255, 238, 229, 248), width: 1),
+                  color: Color.fromARGB(255, 39, 122, 185), width: 1),
               borderRadius: BorderRadius.circular(5)),
-          tileColor: const Color.fromARGB(255, 238, 229, 248),
+          //tileColor: const Color.fromARGB(255, 238, 229, 248),
+          tileColor: const Color.fromARGB(255, 255, 255, 255),
           trailing: SizedBox(
               width: 100,
               child: Row(
@@ -39,11 +40,33 @@ class TarefaItem extends StatelessWidget {
                   // ),
                   IconButton(
                     onPressed: () {
-                      Provider.of<TarefaController>(context, listen: false)
-                          .deleteTarefa(tarefa);
+                      showDialog(
+                        context: context,
+                        builder: (ctx) => AlertDialog(
+                          title: const Text('Excluir Tarefa'),
+                          content:
+                              const Text('Deseja realmente excluir a tarefa?'),
+                          actions: <Widget>[
+                            TextButton(
+                                onPressed: () {
+                                  Provider.of<TarefaController>(context,
+                                          listen: false)
+                                      .deleteTarefa(tarefa);
+                                  Navigator.of(context).pop();
+                                },
+                                child: const Text('Sim')),
+                            TextButton(
+                                onPressed: () {
+                                  Navigator.of(context).pop();
+                                },
+                                child: const Text('Não')),
+                          ],
+                        ),
+                      );
                     },
                     icon: const Icon(Icons.delete),
-                    color: Colors.red,
+                    //color: Colors.red,
+                    color: const Color.fromARGB(255, 183, 31, 20),
                   ),
                 ],
               )),
